Move phylogenetictreenode to shared
