+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/tree/TreeManager.java
@@ -4,9 +4,10 @@ import javafx.animation.Timeline;
 import javafx.scene.control.Button;
 import javafx.scene.input.ScrollEvent;
 import javafx.scene.layout.Pane;
-import nl.tudelft.pl2016gr2.gui.model.IPhylogeneticTreeNode;
+
 import nl.tudelft.pl2016gr2.gui.view.events.GraphicsChangedEvent;
 import nl.tudelft.pl2016gr2.gui.view.selection.SelectionManager;
+import nl.tudelft.pl2016gr2.model.IPhylogeneticTreeNode;
 
 import java.util.ArrayList;
 import java.util.Observer;
