+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/PhylogeneticTreeNode.java
@@ -1,4 +1,4 @@
-package nl.tudelft.pl2016gr2.gui.model;
+package nl.tudelft.pl2016gr2.model;
 
 import java.util.Objects;
